Block Indicator Implemented
Build the cursor() function and now player is able to choose block from a group.
Use the Left and Right Arrow key to switch and Return key to select.

What's next:
Use a test dice button to generate random colors/numbers.
Build two buttons for them and make them as a necessary event for further process.